\name{CLL-package}
\alias{CLL-package}
\alias{CLL}
\docType{package}
\title{
A package for CLL gene expression data
}
\description{
  The CLL package contains the chronic lymphocytic leukemia (CLL) gene
  expression data.  The CLL data had 24 samples that were either classified as
  progressive or stable in regards to disease progression.  The CLL microarray
  data came from Dr. Sabina Chiaretti at Division of Hematology, Department of
  Cellular Biotechnologies and Hematology, University La Sapienza, Rome, Italy
  and Dr. Jerome Ritz at Department of Medicine, Brigham and Women's Hospital,
  Harvard Medical School, Boston, Massachusetts.  
}
\details{
\tabular{ll}{
Package: \tab CLL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-07-10\cr
License: \tab LGPL\cr
}
}
\author{
Elizabeth Whalen

Maintainer: Elizabeth Whalen <ewhalen@hsph.harvard.edu>
}
\keyword{ package }
\seealso{
  \code{\link[<pkg>:<pkg>-ALL]{<pkg>}} 
}
\examples{
  data(sCLLex)
}
